Split runtime retained text edit state
